@@ -78,8 +78,8 @@ describe Delayed::Job, "integration spec", :timeout => 10 do
 
   it "should support run_at" do
     Benchmark.realtime {
-      Delayed::Job.enqueue(TestJobClass.new("job1"), :run_at => Time.now + 0.5)
-      Delayed::Job.enqueue(TestJobClass.new("job2"), :run_at => Time.now + 0.1)
+      Delayed::Job.enqueue(TestJobClass.new("job1"), :run_at => Time.now + 5)
+      Delayed::Job.enqueue(TestJobClass.new("job2"), :run_at => Time.now + 1)
       Delayed::Worker.new.work_off(2).should == [2, 0]
     }.should > 0.1
     @called.should == ["job2", "job1"]
